+++ src/com/epam/task1/service/IMusicService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface IMusicService {
 
-    double getDuration(List<Song> songList);
+    int getDuration(List<Song> songList);
 
-    void getSongByRange(List<Song> songList, double from, double to);
+    void getSongByRange(List<Song> songList, int from, int to);
 
     void sortByStyle(List<Song> songList, Style style);
